favicon work
added favicon for ios safari and macos safari, i wanna see if i can do it for ios chrome. not sure how to check if i can see it on android chrome though. will update.

also resolved warnings in portfolio. i still don't like it but it's the best i got for now.